Update README.md
While I decided to learn to code so I could advance my career, I have been learning a lot of personal lessons as I go on this journey too. I struggle a lot with anxiety, perfectionism, self-criticism, among other things. Its hard for me to start projects unless I know exactly what to do and that I will be successful, and I have a hard time finishing projects if I know they aren't absolutely perfect. I got stuck in that loop on this project particularly. I kept thinking of ways to make it better, and thinking myself in circles trying to figure out the "best" way to perform the task. I didn't want to post and move on from the project, until I thought of every single error that could arise if I actually had a use for this program. I've come to learn that nothing is ever "perfect." I am never going to write an idiot proof code, and there will always improvements that I can make. This is exciting, not scary, and it means I still have a lot to learn. This project represents a lot of growth for me and helped me change something that was anxiety-inducing into something that was enjoyable and energizing. I appreciate the mindful, problem-solving aspects of what I created, even if it isn't perfect, and look forward to applying the new things I will learn.